requirements.txt
Django==3.2.20
psycopg2-binary==2.9.9
WeasyPrint==60.2
django-extensions==3.2.3
django-tailwind==3.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-dotenv==1.0.0
openpyxl==3.1.2
redis==5.0.1
django-redis==5.4.0
django-debug-toolbar==4.2.0
django-compressor==4.4
whitenoise==6.6.0
